% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{build_assessment_column_names}
\alias{build_assessment_column_names}
\title{Build column names from grade and header rows}
\usage{
build_assessment_column_names(grade_row, header_row, sheet_name)
}
\arguments{
\item{grade_row}{Vector with grade labels}

\item{header_row}{Vector with column headers}

\item{sheet_name}{Subject sheet name}
}
\value{
Vector of column names
}
\description{
Build column names from grade and header rows
}
\keyword{internal}
