% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{clean_column_name}
\alias{clean_column_name}
\title{Clean column name}
\usage{
clean_column_name(name)
}
\arguments{
\item{name}{Column name to clean}
}
\value{
Cleaned column name
}
\description{
Clean column name
}
\keyword{internal}
