% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_corp_name}
\alias{clean_corp_name}
\title{Clean corporation/district name}
\usage{
clean_corp_name(x)
}
\arguments{
\item{x}{Character vector of names}
}
\value{
Cleaned character vector
}
\description{
Standardizes corporation names by trimming whitespace and
handling common formatting issues.
}
\keyword{internal}
