% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_school_name}
\alias{clean_school_name}
\title{Clean school name}
\usage{
clean_school_name(x)
}
\arguments{
\item{x}{Character vector of names}
}
\value{
Cleaned character vector
}
\description{
Standardizes school names by trimming whitespace and
handling common formatting issues.
}
\keyword{internal}
