% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{clean_subject_name}
\alias{clean_subject_name}
\title{Clean subject name}
\usage{
clean_subject_name(sheet_name)
}
\arguments{
\item{sheet_name}{Sheet name to clean}
}
\value{
Standardized subject name
}
\description{
Clean subject name
}
\keyword{internal}
