% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{create_empty_assessment_raw}
\alias{create_empty_assessment_raw}
\title{Create empty assessment raw data frame}
\usage{
create_empty_assessment_raw()
}
\value{
Empty data frame with expected columns
}
\description{
Create empty assessment raw data frame
}
\keyword{internal}
