% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_and_merge_corp_data}
\alias{download_and_merge_corp_data}
\title{Download and merge corporation-level enrollment data}
\usage{
download_and_merge_corp_data(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with merged corporation data
}
\description{
Download and merge corporation-level enrollment data
}
\keyword{internal}
