% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_and_merge_school_data}
\alias{download_and_merge_school_data}
\title{Download and merge school-level enrollment data}
\usage{
download_and_merge_school_data(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
Data frame with merged school data
}
\description{
Download and merge school-level enrollment data
}
\keyword{internal}
