% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{download_assessment_file}
\alias{download_assessment_file}
\title{Download a single assessment file}
\usage{
download_assessment_file(end_year, level)
}
\arguments{
\item{end_year}{School year end}

\item{level}{Level: "state", "corporation", or "school"}
}
\value{
Data frame with assessment data
}
\description{
Downloads and parses an assessment Excel file.
}
\keyword{internal}
