% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{find_col}
\alias{find_col}
\title{Find column by multiple possible names}
\usage{
find_col(col_names, possible_names)
}
\arguments{
\item{col_names}{Vector of column names}

\item{possible_names}{Vector of possible column names to match}
}
\value{
First matching column name or NULL
}
\description{
Find column by multiple possible names
}
\keyword{internal}
