% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_idoe_file_urls}
\alias{get_idoe_file_urls}
\title{Get URLs for IDOE enrollment data files}
\usage{
get_idoe_file_urls()
}
\value{
Named list of file URLs
}
\description{
Returns a list of URLs for the various enrollment data files.
IDOE uses Excel files that contain multiple years of data.
}
\keyword{internal}
