% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{get_istep_url}
\alias{get_istep_url}
\title{Get ISTEP+ URL for a given year}
\usage{
get_istep_url(end_year, level, file_type, base_url)
}
\arguments{
\item{end_year}{School year end}

\item{level}{Level of data}

\item{file_type}{Type of file}

\item{base_url}{Base URL for IDOE files}
}
\value{
URL string or NULL
}
\description{
Get ISTEP+ URL for a given year
}
\keyword{internal}
