% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_max_year}
\alias{get_max_year}
\title{Get the maximum available year}
\usage{
get_max_year()
}
\value{
Integer representing most recent available year
}
\description{
Get the maximum available year
}
\keyword{internal}
