% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{get_raw_assessment}
\alias{get_raw_assessment}
\title{Download raw assessment data from IDOE}
\usage{
get_raw_assessment(end_year, level = "all")
}
\arguments{
\item{end_year}{School year end (2014-2019, 2021-2025; no 2020 due to COVID)}

\item{level}{Level of data: "all" (default), "state", "corporation", or "school"}
}
\value{
List with data frames for each level requested
}
\description{
Downloads assessment data from IDOE for the specified year.
Returns raw data from the Excel files without processing.
}
\keyword{internal}
