% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_cache_path}
\alias{get_raw_cache_path}
\title{Get path for raw file cache}
\usage{
get_raw_cache_path(file_type)
}
\arguments{
\item{file_type}{Type of file}
}
\value{
Path to raw cache file
}
\description{
Get path for raw file cache
}
\keyword{internal}
