% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Download raw school directory data from IDOE}
\usage{
get_raw_directory(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
List with corporation, school, and non-public school data frames
}
\description{
Downloads the school directory Excel file from IDOE and reads all sheets.
}
\keyword{internal}
