% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_assessment.R
\name{id_assessment_aggs}
\alias{id_assessment_aggs}
\title{Add aggregation flags to assessment data}
\usage{
id_assessment_aggs(df)
}
\arguments{
\item{df}{Assessment data frame}
}
\value{
Data frame with aggregation flags
}
\description{
Identifies state, corporation, and school level rows based on IDs.
}
\examples{
\dontrun{
assess <- fetch_assessment(2024, tidy = FALSE)
assess <- id_assessment_aggs(assess)
}
}
