% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{process_assessment}
\alias{process_assessment}
\title{Process raw assessment data to standard schema}
\usage{
process_assessment(raw_data, end_year)
}
\arguments{
\item{raw_data}{List with state, corporation, and/or school data frames}

\item{end_year}{School year end}
}
\value{
Data frame with processed assessment data
}
\description{
Takes raw assessment data from IDOE and processes it into a consistent
schema suitable for analysis.
}
\keyword{internal}
