% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{process_corp_directory}
\alias{process_corp_directory}
\title{Process corporation directory data}
\usage{
process_corp_directory(df)
}
\arguments{
\item{df}{Raw corporation data frame}
}
\value{
Processed corporation data frame
}
\description{
Process corporation directory data
}
\keyword{internal}
