% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_corp_enr}
\alias{process_corp_enr}
\title{Process corporation-level enrollment data}
\usage{
process_corp_enr(df, end_year)
}
\arguments{
\item{df}{Raw corporation data frame}

\item{end_year}{School year end}
}
\value{
Processed corporation data frame
}
\description{
Process corporation-level enrollment data
}
\keyword{internal}
