% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{process_np_school_directory}
\alias{process_np_school_directory}
\title{Process non-public school directory data}
\usage{
process_np_school_directory(df)
}
\arguments{
\item{df}{Raw non-public school data frame}
}
\value{
Processed non-public school data frame
}
\description{
Process non-public school directory data
}
\keyword{internal}
