% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{process_school_directory}
\alias{process_school_directory}
\title{Process public school directory data}
\usage{
process_school_directory(df)
}
\arguments{
\item{df}{Raw school data frame}
}
\value{
Processed school data frame
}
\description{
Process public school directory data
}
\keyword{internal}
