% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{read_assessment_excel}
\alias{read_assessment_excel}
\title{Read assessment data from Excel file}
\usage{
read_assessment_excel(filepath, end_year, level)
}
\arguments{
\item{filepath}{Path to Excel file}

\item{end_year}{School year end}

\item{level}{Data level}
}
\value{
Data frame with assessment data
}
\description{
Parses IDOE assessment Excel files. These files have multi-row headers
and multiple sheets for different subjects.
}
\keyword{internal}
